+++ views/src/components/diary/DiaryReadComponent.js
@@ -51,24 +51,31 @@ const DiaryReadComponent = () => {
   }, []);
 
   const deleteOnClick = () => {
-    deleteDiaryC(diaryParam.id)
-      .then((result) => {
-        console.log("DiaryReadComponent.js deleteOnClick() then() => ", result);
-        if (result.payload.result === "success") {
-          alert(diaryParam.dateobject + "일자의 다이어리가 삭제되었습니다.");
+    if (window.confirm("해당 다이어리를 삭제하십니까?")) {
+      deleteDiaryC(diaryParam.id)
+        .then((result) => {
+          console.log(
+            "DiaryReadComponent.js deleteOnClick() then() => ",
+            result
+          );
+          if (result.payload.result === "success") {
+            alert(diaryParam.dateobject + "일자의 다이어리가 삭제되었습니다.");
 
-          moveToDiaryList();
-        }
-        if (result.payload.error === "error") {
+            moveToDiaryList();
+          }
+          if (result.payload.error === "error") {
+            alert(
+              diaryParam.dateobject + "일자의 다이어리 삭제를 실패했습니다."
+            );
+            moveToDiaryRead(diaryParam.id);
+          }
+        })
+        .catch((err) => {
+          console.log("DiaryReadComponent.js deleteOnClick() catch() => ", err);
           alert(diaryParam.dateobject + "일자의 다이어리 삭제를 실패했습니다.");
           moveToDiaryRead(diaryParam.id);
-        }
-      })
-      .catch((err) => {
-        console.log("DiaryReadComponent.js deleteOnClick() catch() => ", err);
-        alert(diaryParam.dateobject + "일자의 다이어리 삭제를 실패했습니다.");
-        moveToDiaryRead(diaryParam.id);
-      });
+        });
+    } //window.confirm
   };
 
   return (
@@ -106,7 +113,7 @@ const DiaryReadComponent = () => {
           {diaryParam.picture ? (
             <Image
               //src={`http://localhost:8001/img/${diaryParam.picture}`}
-              src={`https://picdiary-bucket.s3.ap-northeast-2.amazonaws.com/${diaryParam.picture}`}
+              src={`https://picdiary-bucket.s3.ap-northeast-2.amazonaws.com/${diaryParam.path}`}
               rounded
             />
           ) : (
